requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
blinker==1.7.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-mail==1.4.1
idna==3.6
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
Pillow==10.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
